refactor(dashboards): inline the markdown format tools with block controls
Lift the TipTap editor into TextWidgetCard so its format tools (bold/italic/H2/
lists) render inline in the block toolbar while editing, alongside Arrange/
Expand/Save/Delete. The editor also renders the note read-only when not editing
(replacing the separate react-markdown view), so one instance drives both
states. Removes the standalone MarkdownEditor component.
